add github workflow when opening a pr

+++ internal/model/order_test.go
@@ -6,14 +6,13 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-
 func TestNewOrder(t *testing.T) {
 	items := []OrderItem{
 		{
 			ID:        1,
 			OrderID:   10,
 			ProductID: 1,
-			Product: nil,
+			Product:   nil,
 			Qty:       1,
 			UnitPrice: 150.0,
 		},
@@ -21,15 +20,15 @@ func TestNewOrder(t *testing.T) {
 
 	payments := []Payment{
 		{
-			ID: 1,
-			OrderID: 1,
-			Provider: "stripe",
-			AmountCents: 150.0,
-			Method: "credit_card",
-			Currency: "USD",
-			Status: "pending",
+			ID:                    1,
+			OrderID:               1,
+			Provider:              "stripe",
+			AmountCents:           150.0,
+			Method:                "credit_card",
+			Currency:              "USD",
+			Status:                "pending",
 			StripePaymentIntentID: nil,
-			StripeChargeID: nil,
+			StripeChargeID:        nil,
 		},
 	}
 
@@ -37,26 +36,26 @@ func TestNewOrder(t *testing.T) {
 
 	assert.NoError(t, err)
 	assert.NotNil(t, order)
-	
+
 	// Assert basic order fields
 	assert.Equal(t, 1, order.UserID)
 	assert.Equal(t, "pending", order.Status)
 	assert.Equal(t, 150.0, order.TotalPrice)
 	assert.Equal(t, "USD", order.Currency)
-	
+
 	// Assert order has correct number of items and payments
 	assert.Len(t, order.Items, 1)
 	assert.Len(t, order.Payments, 1)
-	
+
 	// Assert item details
 	assert.Equal(t, 1, order.Items[0].ProductID)
 	assert.Equal(t, 1, order.Items[0].Qty)
 	assert.Equal(t, 150.0, order.Items[0].UnitPrice)
-	
+
 	// Assert payment details
 	assert.Equal(t, "stripe", order.Payments[0].Provider)
 	assert.Equal(t, "credit_card", order.Payments[0].Method)
-	assert.Equal(t, 150.0, order.Payments[0].AmountCents)
+	assert.Equal(t, int64(150.0), order.Payments[0].AmountCents)
 	assert.Equal(t, "USD", order.Payments[0].Currency)
 	assert.Equal(t, "pending", order.Payments[0].Status)
 }
@@ -70,7 +69,7 @@ func TestNewOrderWithMissingUserID(t *testing.T) {
 	}
 
 	order, err := NewOrder(0, items, "pending", 150.0, "USD", payments)
-	
+
 	assert.Error(t, err)
 	assert.Nil(t, order)
 	assert.Contains(t, err.Error(), "user ID is required")
@@ -82,7 +81,7 @@ func TestNewOrderWithNoItems(t *testing.T) {
 	}
 
 	order, err := NewOrder(1, []OrderItem{}, "pending", 0.0, "USD", payments)
-	
+
 	assert.Error(t, err)
 	assert.Nil(t, order)
 	assert.Contains(t, err.Error(), "order must have at least one item")
@@ -97,7 +96,7 @@ func TestNewOrderWithNegativeQuantity(t *testing.T) {
 	}
 
 	order, err := NewOrder(1, items, "pending", 150.0, "USD", payments)
-	
+
 	assert.Error(t, err)
 	assert.Nil(t, order)
 	assert.Contains(t, err.Error(), "item quantity must be positive")
@@ -112,7 +111,7 @@ func TestNewOrderWithZeroQuantity(t *testing.T) {
 	}
 
 	order, err := NewOrder(1, items, "pending", 150.0, "USD", payments)
-	
+
 	assert.Error(t, err)
 	assert.Nil(t, order)
 	assert.Contains(t, err.Error(), "item quantity must be positive")
@@ -157,26 +156,26 @@ func TestNewOrderTableDriven(t *testing.T) {
 			wantErr:    true,
 		},
 		{
-			name:		"quantity must be positive",
+			name:   "quantity must be positive",
 			userID: 1,
 			items: []OrderItem{
 				{ProductID: 1, Qty: -1, UnitPrice: 150.0},
 			},
-			status: "pending",
+			status:     "pending",
 			totalPrice: 150.0,
-			currency: "USD",
-			payments: validPayments,
-			wantErr: true,
+			currency:   "USD",
+			payments:   validPayments,
+			wantErr:    true,
 		},
 		{
-			name:		"order must have at least one item",
-			userID: 1,
-			items: []OrderItem{},
-			status: "pending",
+			name:       "order must have at least one item",
+			userID:     1,
+			items:      []OrderItem{},
+			status:     "pending",
 			totalPrice: 150.0,
-			currency: "USD",
-			payments: validPayments,
-			wantErr: true,
+			currency:   "USD",
+			payments:   validPayments,
+			wantErr:    true,
 		},
 	}
